+++ phoenix/tests/test_utils.py
@@ -2,6 +2,7 @@ from datetime import datetime, timedelta
 
 from phoenix import utils
 
+
 def test_time_ago_in_words():
     assert utils.time_ago_in_words(datetime.now() - timedelta(minutes=1)) == '1 minute ago'
     assert utils.time_ago_in_words(datetime.now() - timedelta(hours=1, minutes=7)) == '1 hour ago'
@@ -10,16 +11,18 @@ def test_time_ago_in_words():
     assert utils.time_ago_in_words(datetime.now() - timedelta(days=31, hours=1)) == '1 month ago'
     assert utils.time_ago_in_words(datetime.now() - timedelta(days=366, hours=1)) == '1 year ago'
     assert utils.time_ago_in_words('nothing') == '???'
-    
+
+
 def test_make_tags():
     assert utils.make_tags("main,dev") == ['dev', 'main']
     assert utils.make_tags("main,dev,Testing") == ['dev', 'main', 'testing']
-    assert utils.make_tags("public, dev,Testing , Next Generation, ,,; ") == [';', 'dev', 'next generation', 'public', 'testing']
+    assert utils.make_tags("public, dev,Testing , Next Generation, ,,; ") ==\
+        [';', 'dev', 'next generation', 'public', 'testing']
     assert utils.make_tags("public, dev, test, test") == ['dev', 'public', 'test']
     assert utils.make_tags("public, dev, private, all") == ['dev', 'public']
-    
+
+
 def test_format_tags():
     assert utils.format_tags(['public', 'dev']) == 'public, dev'
     assert utils.format_tags(None) == ''
     assert utils.format_tags([]) == ''
-