+++ phoenix/wizard/views/esgflogin.py
@@ -16,9 +16,10 @@ class ESGFLoginSchema(colander.MappingSchema):
     openid = colander.SchemaNode(
         colander.String(),
         title="OpenID",
-        description="Type your OpenID from your ESGF provider. For example: https://esgf-data.dkrz.de/esgf-idp/openid/username",
+        description="Type your OpenID from your ESGF provider.\
+         For example: https://esgf-data.dkrz.de/esgf-idp/openid/username",
         validator=colander.url
-        )
+    )
     password = colander.SchemaNode(
         colander.String(),
         title='Password',
@@ -52,16 +53,15 @@ class ESGFLogin(Wizard):
         super(ESGFLogin, self).success(appstruct)
 
         self.wizard_state.set('password', appstruct.get('password'))
-        result = esgf_logon.delay(authenticated_userid(self.request), self.request.wps.url, 
-                            appstruct.get('openid'),
-                            appstruct.get('password'))
+        result = esgf_logon.delay(authenticated_userid(self.request), self.request.wps.url,
+                                  appstruct.get('openid'),
+                                  appstruct.get('password'))
         self.session['task_id'] = result.id
 
     def next_success(self, appstruct):
         self.success(appstruct)
         return HTTPFound(location=self.request.route_path('wizard_loading'))
 
-    @view_config(renderer='json', route_name='wizard_check_logon')
     def check_logon(self):
         status = 'running'
         result = task_result(self.session.get('task_id'))
@@ -69,7 +69,6 @@ class ESGFLogin(Wizard):
             status = 'ready'
         return dict(status=status)
 
-    @view_config(route_name='wizard_loading', renderer='../templates/wizard/loading.pt')
     def loading(self):
         result = task_result(self.session.get('task_id'))
         if result.ready():
@@ -80,7 +79,6 @@ class ESGFLogin(Wizard):
                 self.session.flash('ESGF logon failed.', queue='danger')
                 return HTTPFound(location=self.request.route_path(self.name))
         return {}
-        
-    @view_config(route_name='wizard_esgf_login', renderer='../templates/wizard/default.pt')
+
     def view(self):
         return super(ESGFLogin, self).view()